lib1582.c drop redundant cast

+++ tests/libtest/lib1582.c
@@ -43,7 +43,7 @@ static CURLcode test_lib1582(const char *URL)
   easy_setopt(curl, CURLOPT_HEADER, 1L);
   easy_setopt(curl, CURLOPT_VERBOSE, 1L);
   easy_setopt(curl, CURLOPT_URL, URL);
-  easy_setopt(curl, CURLOPT_HTTPAUTH, (long)CURLAUTH_NEGOTIATE);
+  easy_setopt(curl, CURLOPT_HTTPAUTH, CURLAUTH_NEGOTIATE);
   easy_setopt(curl, CURLOPT_USERPWD, ":");
   easy_setopt(curl, CURLOPT_SSL_VERIFYPEER, 0L);
   easy_setopt(curl, CURLOPT_SSL_VERIFYHOST, 0L);
